move checkAccessionMatch function to events.js, add loading animations after final submit modal, fix promises on modals
