+++ src/pages/specimen.js
@@ -1,5 +1,5 @@
-import { addEventBarCodeScanner, collectionSettings, generateBarCode, getWorkflow, removeActiveClass, siteLocations, visitType, getCheckedInVisit, getSiteAcronym, numericInputValidator, getSiteCode, searchSpecimen, collectionInputValidator, addSelectionEventListener, autoTabInputField } from "./../shared.js";
-import { addEventSpecimenLinkForm, addEventClinicalSpecimenLinkForm, addEventClinicalSpecimenLinkForm2, addEventNavBarParticipantCheckIn, addEventBackToSearch } from "./../events.js";
+import { addEventBarCodeScanner, collectionSettings, generateBarCode, getWorkflow, removeActiveClass, siteLocations, visitType, getCheckedInVisit, getSiteAcronym, numericInputValidator, getSiteCode, searchSpecimen, collectionInputValidator, addSelectionEventListener, autoTabInputField, errorMessage, removeAllErrors, removeSingleError } from "./../shared.js";
+import { addEventSpecimenLinkForm, addEventClinicalSpecimenLinkForm, addEventClinicalSpecimenLinkForm2, addEventNavBarParticipantCheckIn, addEventBackToSearch, checkAccessionMatch } from "./../events.js";
 import { masterSpecimenIDRequirement } from "../tubeValidation.js";
 import { conceptIds } from '../fieldToConceptIdMapping.js';
 
@@ -151,7 +151,6 @@ export const specimenTemplate = async (data, formData) => {
             </div>
             <div class="form-group row">
                 <input autocomplete="off" type="text" class="form-control col-md-5 offset-4" placeholder="Re-Enter (scan/type) in Accession ID from Blood Tube" id="accessionID2" maxlength="11"/>
-                <div id="accessionId2Error" class="invalid-feedback col-md-5 offset-4 form-helper-text" style="padding-left:0; "></div>
             </div>
             </br>
             <div class="form-group row">
@@ -162,7 +161,6 @@ export const specimenTemplate = async (data, formData) => {
             </div>
             <div class="form-group row">
                 <input autocomplete="off" type="text" class="form-control col-md-5 offset-4" placeholder="Re-Enter (scan/type) in Accession ID from Urine Tube" id="accessionID4" maxlength="11"/>
-                <div id="accessionId4Error" class="invalid-feedback col-md-5 offset-4 form-helper-text" style="padding-left:0; "></div>
             </div>
             
         <div class="form-group row">
@@ -200,99 +198,12 @@ export const specimenTemplate = async (data, formData) => {
         document.getElementById('accessionID4').onpaste = e => e.preventDefault();
 
         numericInputValidator(['accessionID1', 'accessionID2', 'accessionID3', 'accessionID4']);
-
-        // add form validation on user input for accession IDs 3 and 4
-        checkAccessionMatch(['accessionID1', 'accessionID2', 'accessionID3', 'accessionID4']);
+        checkAccessionMatch();
         
-        addEventClinicalSpecimenLinkForm(formData);
+        addEventClinicalSpecimenLinkForm(data, formData);
     }
 
     generateBarCode('connectIdBarCode', data.Connect_ID);
     addEventBackToSearch('navBarSearch');
     addEventNavBarParticipantCheckIn();
 }
-
-/**
- * Adds event listeners to the form fields to make sure accession inputs are checked and have matching values
-*/
-const checkAccessionMatch = () => { 
-    const accessionId1 = document.getElementById('accessionID1');
-    const accessionId2 = document.getElementById('accessionID2');
-    const accessionId3 = document.getElementById('accessionID3');
-    const accessionId4 = document.getElementById('accessionID4');
-    const accessionId2Error = document.getElementById('accessionId2Error');
-    const accessionId4Error = document.getElementById('accessionId4Error');
-        
-    accessionId2.addEventListener('blur', () => { 
-        if (accessionId1.value !== accessionId2.value) {
-            addAccessionError(accessionId1, accessionId2, accessionId2Error, 'blood');
-        } else {
-            removeAccessionError(accessionId1, accessionId2, accessionId2Error);
-        }
-    });
-
-    accessionId4.addEventListener('blur', () => { 
-        if (accessionId4.value !== accessionId3.value) {
-            addAccessionError(accessionId3, accessionId4, accessionId4Error, 'urine');
-        } else {
-            removeAccessionError(accessionId3, accessionId4, accessionId4Error);
-        }
-    });
-
-    // add blur event handler logic the accession Id inputs for 1 and 2, when user retypes bloor or urine (not reenter input fields)
-    accessionId1.addEventListener('blur', () => {
-        if (accessionId1.classList.contains('is-invalid')) {
-            if (accessionId1.value === accessionId2.value) {
-                removeAccessionError(accessionId1, accessionId2, accessionId2Error);
-            }
-        }
-
-        if (accessionId1.value.length > 0 && 
-            accessionId2.value.length > 0 && 
-            accessionId1.value !== accessionId2.value) {
-            addAccessionError(accessionId1, accessionId2, accessionId2Error, 'blood');
-        }
-    });
-
-    accessionId3.addEventListener('blur', () => {
-        if (accessionId3.classList.contains('is-invalid')) {
-            if (accessionId3.value === accessionId4.value) {
-                removeAccessionError(accessionId3, accessionId4, accessionId4Error);
-            }
-        }
-        if (accessionId3.value.length > 0 && 
-            accessionId4.value.length > 0 && 
-            accessionId3.value !== accessionId4.value) {
-            addAccessionError(accessionId3, accessionId4, accessionId4Error, 'urine'); 
-        }
-    });
-};
-
-/**
- * Adds error class ".is-invalid" and error message to the accession input field
- * @param {Element} accessionInput1 - doc element of input  (Ex. accessionID1)
- * @param {Element} accessionInput2 - doc element of input  (Ex. accessionID2)
- * @param {Element} accessionError - accession for element error text, where the error message will display
- * @param {string} accessionType - blood or urine
- */
-const addAccessionError = (accessionInput1, accessionInput2, accessionError, accessionType) => { 
-    accessionInput1.classList.add('is-invalid');
-    accessionInput2.classList.add('is-invalid');
-
-    if (accessionError) {
-        accessionError.textContent = `The ${accessionType} accession IDs do not match.`;
-    }
-};
-
-/**
- * Removes error class ".is-invalid" and error message from the accession input field
- * @param {Element} accessionInput1 - doc element of input  (Ex. accessionID1)
- * @param {Element} accessionInput2 - doc element of input  (Ex. accessionID2)
- * @param {Element} accessionError - accession for element Error
-*/
-const removeAccessionError = (accessionInput1, accessionInput2, accessionError) => { 
-    accessionInput1.classList.remove('is-invalid');
-    accessionInput2.classList.remove('is-invalid');
-
-    if (accessionError) accessionError.textContent = '';
-};
